Move mouse cursor handling into own class: MouseShapeController.
This enables moving the CameraRotationController instance into
the DemoScene where it now gets a reference to the camera.
The MouseShapeController can take its place in the main function
and get the window.
This changes not only enables getting the dependencies right,
but also provides a reusable class for other dragging cases.